- Feat: Java 11.
BREAKING CHANGE: Java 8 a Java 11.

+++ src/main/java/motor/core/graphics/Imagen.java
@@ -111,14 +111,14 @@ public final class Imagen {
      * @return el ancho
      */
     public int getAncho() {
-        return (int) anchoEscalado;
+        return anchoEscalado;
     }
 
     /**
      * @return el alto
      */
     public int getAlto() {
-        return (int) altoEscalado;
+        return altoEscalado;
     }
 
     /**
